modified:   float.cpp 	new file:   long.cpp 	new file:   long_double.cpp 	modified:   run 	modified:   long.cpp 	modified:   run

+++ long_double.cpp
@@ -3,7 +3,7 @@
 int main() {
     int n=1;
     int repeats=100000000;
-    float pi=1;
+    long double pi=1;
     while (n<repeats) {
         pi=pi * 2*n/(2*n-1) * 2*n/(2*n+1);
         n++;
